requirements.txt
# FastAPI and its server
fastapi
uvicorn[standard]

# For handling file uploads and forms
python-multipart

# For HTML templating (FastAPI doesn't include it by default)
jinja2

# Data handling
pandas

# PDF processing
pdfplumber

# Machine Learning and similarity calculation
scikit-learn
numpy

# Google AI and LangChain
langchain-google-genai
langchain-core
google-generativeai
python-dotenv

# For SecretStr
pydantic

# for rendering markdown in UI
markdown2